+++ torchmetrics/__init__.py
@@ -13,7 +13,15 @@ _PROJECT_ROOT = os.path.dirname(_PACKAGE_ROOT)
 
 from torchmetrics import functional  # noqa: E402
 from torchmetrics.aggregation import CatMetric, MaxMetric, MeanMetric, MinMetric, SumMetric  # noqa: E402
-from torchmetrics.audio import PIT, SDR, SI_SDR, SI_SNR, SNR  # noqa: E402
+from torchmetrics.audio import (  # noqa: E402
+    PIT,
+    SDR,
+    SI_SDR,
+    SI_SNR,
+    SNR,
+    ScaleInvariantSignalDistortionRatio,
+    SignalDistortionRatio,
+)
 from torchmetrics.classification import (  # noqa: E402, F401
     AUC,
     AUROC,
@@ -143,6 +151,8 @@ __all__ = [
     "ROC",
     "SacreBLEUScore",
     "SDR",
+    "SignalDistortionRatio",
+    "ScaleInvariantSignalDistortionRatio",
     "SI_SDR",
     "SI_SNR",
     "SNR",